requirements.txt
numpy>=1.20.0
Pillow>=8.0.0
pandas>=1.3.0
openpyxl>=3.0.0
